requirements.txt
Jinja2==3.0.3
PyYAML==6.0